requirements.txt
confluent-kafka
pdfplumber
google-cloud-storage